+++ src/main/java/com/codals/greating/diy/controller/DiyRestController.java
@@ -1,0 +1,30 @@
+package com.codals.greating.diy.controller;
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/mealdiy")
+public class DiyRestController {
+
+    @PostMapping
+    public String search(@RequestParam(required = false) String category,
+                         @RequestParam(required = false) String classification,
+                         @RequestParam(required = false) boolean includeRice,
+                         @RequestParam(required = false) boolean includeSoup) {
+        /**
+         * diy 메인 > 돋보기, 선택완료 버튼 (Ajax)
+         */
+        return "diy/diy-main";
+    }
+
+    @PostMapping("/new")
+    public String create() {
+        /**
+         * DIY 식단 저장, 상세보기 리다이렉트 (Ajax)
+         */
+        return "redirect:/mealdiy/" + "1";
+    }
+}
